+++ lib/graphql/schema/discussions/comment.ex
@@ -55,7 +55,7 @@ defmodule Mobilizon.GraphQL.Schema.Discussions.CommentType do
       description: "Whether this comment needs to be announced to participants"
     )
 
-    field(:language, non_null(:string), description: "The comment language")
+    field(:language, :string, description: "The comment language")
   end
 
   @desc "The list of visibility options for a comment"