+++ lib/graphql/schema/post.ex
@@ -20,7 +20,7 @@ defmodule Mobilizon.GraphQL.Schema.PostType do
     field(:publish_at, :datetime, description: "When the post was published")
     field(:inserted_at, :datetime, description: "The post's creation date")
     field(:updated_at, :datetime, description: "The post's last update date")
-    field(:language, non_null(:string), description: "The post language")
+    field(:language, :string, description: "The post language")
 
     field(:tags, list_of(:tag),
       resolve: &Tag.list_tags_for_post/3,